perf: change list box item to memoized component
